+++ src/commands/settings/mod.rs
@@ -1,16 +1,14 @@
 use serenity::framework::standard::macros::group;
 
 mod default;
-mod join_msg;
-mod leave_msg;
 mod list;
 mod mod_log;
 mod mute;
 
-use self::{default::*, join_msg::*, leave_msg::*, list::*, mod_log::*, mute::*};
+use self::{default::*, list::*, mod_log::*, mute::*};
 
 #[group]
-#[commands(list, modlog, mute, joinmsg, leavemsg)]
+#[commands(list, modlog, mute)]
 #[description("Guild settings, requires MANAGE_GUILD permissions")]
 #[prefix("settings")]
 #[only_in("guild")]