+++ src/model/sql/mod.rs
@@ -1,11 +1,13 @@
 pub mod guild;
 pub mod guild_roles;
+pub mod guild_setting;
 pub mod mod_log;
 pub mod mute;
 
 pub use self::{
     guild::{GuildConfig, GuildConfigDb},
     guild_roles::GuildRoles,
+    guild_setting::{GuildSetting, GuildSettingAction},
     mod_log::{ModLogEntry, ModLogEntryDb},
     mute::{delete_mute, Mute, MuteDb},
 };