+++ src/utils/mod.rs
@@ -1,4 +1,2 @@
 pub mod duration;
 pub mod user;
-#[macro_use]
-pub mod settings;